requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
xgboost==0.81
tqdm
tabulate
Keras
pandas
tables==3.4.4
joblib
tensorflow
sklearn
